Modification of yaml-config implementation
